+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorEnumValidationTest.kt
@@ -49,7 +49,7 @@ class CodeGeneratorEnumValidationTest {
 
     @Test fun `should generate correct code for string enum validations in Java`() {
         val input = File("src/test/resources/test-enum-validation.schema.json")
-        val codeGenerator = CodeGenerator(templates = "java", suffix = "java")
+        val codeGenerator = CodeGenerator(TargetLanguage.JAVA)
         val stringWriter = StringWriter()
         codeGenerator.basePackageName = "com.example"
         codeGenerator.outputResolver = outputCapture(TargetFileName("TestEnumValidation", "java", dirs), stringWriter)