+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorNestedClassTest.kt
@@ -49,7 +49,7 @@ class CodeGeneratorNestedClassTest {
 
     @Test fun `should output deeply nested class in Java`() {
         val input = File("src/test/resources/test-nested-object.schema.json")
-        val codeGenerator = CodeGenerator(templates = "java", suffix = "java")
+        val codeGenerator = CodeGenerator(TargetLanguage.JAVA)
         val stringWriter = StringWriter()
         codeGenerator.basePackageName = "com.example"
         codeGenerator.outputResolver = outputCapture(TargetFileName("TestNestedObject", "java", dirs), stringWriter)